Config Update
- Now default location for config file is `home/.config/cook/cook.yaml`
- Separated Config File
